# test(benchmarks) Add the `InvokeFunction` benchmark